frankenapp/orchestrator/tests/test_health.py
"""Basic health check tests for the orchestrator."""

import pytest
from httpx import AsyncClient, ASGITransport

from app.main import app


@pytest.mark.asyncio
async def test_health_endpoint():
    """The /health endpoint should return 200 with status ok."""
    async with AsyncClient(
        transport=ASGITransport(app=app), base_url="http://test"
    ) as client:
        response = await client.get("/health")
    assert response.status_code == 200
    data = response.json()
    assert data["status"] == "ok"
    assert "koboldcpp" in data
    assert "comfyui" in data


@pytest.mark.asyncio
async def test_docs_available():
    """OpenAPI docs endpoint should be accessible."""
    async with AsyncClient(
        transport=ASGITransport(app=app), base_url="http://test"
    ) as client:
        response = await client.get("/docs")
    assert response.status_code == 200
