frankenapp/orchestrator/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
websockets>=12.0
pydantic-settings>=2.1.0
aiosqlite>=0.19.0
Pillow>=10.2.0
python-multipart>=0.0.6
tiktoken>=0.6.0
